[Refactor] components 분리
