+++ src/data/posts.js
@@ -1,0 +1,37 @@
+export const posts = [
+  {
+    id: 1,
+    author: "홍길동",
+    profileImage: "https://placehold.co/40",
+    content: "오늘 날씨가 정말 좋네요. 다들 좋은 하루 보내세요!",
+    timeAgo: "5분 전",
+  },
+  {
+    id: 2,
+    author: "김철수",
+    profileImage: "https://placehold.co/40",
+    content: "새로운 프로젝트를 시작했어요. 열심히 코딩 중입니다!",
+    timeAgo: "15분 전",
+  },
+  {
+    id: 3,
+    author: "이영희",
+    profileImage: "https://placehold.co/40",
+    content: "오늘 점심 메뉴 추천 받습니다. 뭐가 좋을까요?",
+    timeAgo: "30분 전",
+  },
+  {
+    id: 4,
+    author: "박민수",
+    profileImage: "https://placehold.co/40",
+    content: "주말에 등산 가실 분 계신가요? 함께 가요!",
+    timeAgo: "1시간 전",
+  },
+  {
+    id: 5,
+    author: "정수연",
+    profileImage: "https://placehold.co/40",
+    content: "새로 나온 영화 재미있대요. 같이 보러 갈 사람?",
+    timeAgo: "2시간 전",
+  },
+];
